services/ai-service/requirements.txt
fastapi
uvicorn
pdfplumber
requests
numpy
pika
minio
qdrant-client
python-multipart
python-dotenv
pymongo